backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
groq
google-generativeai
requests
python-multipart